+++ resolwe_bio/processes/reads_processing/markduplicates.py
@@ -24,7 +24,7 @@ class MarkDuplicates(Process):
     slug = "markduplicates"
     name = "MarkDuplicates"
     process_type = "data:alignment:bam:markduplicate:"
-    version = "1.6.0"
+    version = "1.7.0"
     category = "BAM processing"
     scheduling_class = SchedulingClass.BATCH
     entity = {"type": "sample"}
@@ -82,22 +82,6 @@ class MarkDuplicates(Process):
             default="",
         )
 
-        class BigWigOptions:
-            """Options for calculating BigWig."""
-
-            bigwig_binsize = IntegerField(
-                label="BigWig bin size",
-                description="Size of the bins, in bases, for the output of the "
-                "bigwig/bedgraph file. Default is 50.",
-                default=50,
-            )
-            bigwig_timeout = IntegerField(
-                label="BigWig timeout",
-                description="Number of seconds before creation of BigWig timeouts. "
-                "Default is after 480 seconds (8 minutes).",
-                default=480,
-            )
-
         class Advanced:
             """Advanced options."""
 
@@ -112,8 +96,6 @@ class MarkDuplicates(Process):
                 description="Set the maximum Java heap size (in GB).",
             )
 
-        bigwig_opts = GroupField(BigWigOptions, label="BigWig options")
-
         advanced = GroupField(Advanced, label="Advanced options")
 
     class Output:
@@ -122,7 +104,6 @@ class MarkDuplicates(Process):
         bam = FileField(label="Marked duplicates BAM file")
         bai = FileField(label="Index of marked duplicates BAM file")
         stats = FileField(label="Alignment statistics")
-        bigwig = FileField(label="BigWig file", required=False)
         species = StringField(label="Species")
         build = StringField(label="Build")
         metrics_file = FileField(label="Metrics from MarkDuplicate process")
@@ -210,25 +191,6 @@ class MarkDuplicates(Process):
 
         stats = f"{bam}_stats.txt"
         (Cmd["samtools"]["flagstat"][f"{bam}"] > stats)()
-
-        self.progress(0.8)
-
-        btb_inputs = [
-            bam,
-            species,
-            self.requirements.resources.cores,
-            "deeptools",
-            inputs.bigwig_opts.bigwig_binsize,
-            inputs.bigwig_opts.bigwig_timeout,
-        ]
-
-        Cmd["bamtobigwig.sh"](btb_inputs)
-
-        bigwig = bam[:-4] + ".bw"
-        if not os.path.exists(bigwig):
-            self.info("BigWig file not calculated.")
-        else:
-            outputs.bigwig = bigwig
 
         self.progress(0.9)